requirements.txt
PyGithub==2.4.0
pylate==1.0.0
PyPDF2==3.0.1
python-dotenv==1.0.1
